refactor: consolidate scene operations into a single scenes WKC component
Replace the multi-module shape (probe-scene.ts + asset-reuse.ts function
exports) with one importable component that's responsible for the scene
pre-Unity pipeline. The orchestrator and executeConversion / executeTriagePass
now reach for these operations through components.scenes.* instead of
importing free functions across two modules.

logic/scenes/{component,types,index}.ts:
- IScenesComponent interface with probe, uploadFastPathResult, checkAssetCache,
  computePerAssetDigests, purgeCachedBundlesFromOutput, shouldIgnoreConversion,
  getCdnBucket, manifestKeyForEntity, uploadEntityManifest,
  uploadSceneSourceFilesToCDN.
- createScenesComponent factory takes Pick<AppComponents, ...> deps and returns
  the interface. Probe + fast-path-upload bodies live inline in the factory
  (these were the entirety of probe-scene.ts, now deleted); the heavier S3 /
  GLB operations delegate to free-function implementations in asset-reuse.ts.
  asset-reuse.ts kept in place — its 1500-line unit test suite stays stable.

src/types.ts: AppComponents gains `scenes: IScenesComponent`.
src/components.ts: instantiates the scenes component and threads it through.
conversion-orchestrator: 'scenes' added to its Pick<>.
conversion-task.ts: every probe / cdn-bucket / manifest-key / upload /
purge call rewritten as components.scenes.*.
probe-scene.ts: deleted.

Test harnesses that construct components manually now build a scenes
instance too. Three orchestrator-unit-style harnesses use `scenes: {} as any`
because their executeConversion / executeTriagePass calls are jest.mock'd at
module scope and never reach into the scenes methods.

424 tests pass, lint + tsc clean. No behaviour change — pure structural
refactor consolidating the public surface.
