+++ consumer-server/src/logic/scenes/index.ts
@@ -1,0 +1,10 @@
+export { createScenesComponent } from './component'
+export type {
+  CheckAssetCacheArgs,
+  ComputePerAssetDigestsOptions,
+  IScenesComponent,
+  Manifest,
+  ProbeArgs,
+  ProbeOutcome,
+  UploadFastPathArgs
+} from './types'
