+++ consumer-server/test/integration/triage-then-conversion.spec.ts
@@ -84,9 +84,11 @@ describe('when FAST_PATH_TRIAGE_ENABLED is true', () => {
       // Stubs — these tests jest.mock executeConversion / executeLODConversion
       // / executeTriagePass at module scope, so the orchestrator never actually
       // dispatches into the catalyst or unity-runner methods. Empty fakes
-      // satisfy the type contract.
+      // satisfy the type contract. The scenes component is similarly unused on
+      // these paths; pass an empty object cast to satisfy the Pick.
       catalyst: { getActiveEntity: jest.fn(), getEntities: jest.fn() },
-      unityRunner: { runConversion: jest.fn(), runLodsConversion: jest.fn() }
+      unityRunner: { runConversion: jest.fn(), runLodsConversion: jest.fn() },
+      scenes: {} as any
     })
 
     const components = {
@@ -407,9 +409,11 @@ describe('when FAST_PATH_TRIAGE_ENABLED is unset (default off)', () => {
       // Stubs — these tests jest.mock executeConversion / executeLODConversion
       // / executeTriagePass at module scope, so the orchestrator never actually
       // dispatches into the catalyst or unity-runner methods. Empty fakes
-      // satisfy the type contract.
+      // satisfy the type contract. The scenes component is similarly unused on
+      // these paths; pass an empty object cast to satisfy the Pick.
       catalyst: { getActiveEntity: jest.fn(), getEntities: jest.fn() },
-      unityRunner: { runConversion: jest.fn(), runLodsConversion: jest.fn() }
+      unityRunner: { runConversion: jest.fn(), runLodsConversion: jest.fn() },
+      scenes: {} as any
     })
 
     const components = {
